// lib/enums.ts
export enum AccountCategoryType {
  ASSET = "ASSET",
  LIABILITY = "LIABILITY",
}

export enum AccountType {
  TRACKED = "TRACKED",
  VALUATED = "VALUATED",
}

export enum AccountUnitKind {
  CURRENCY = "CURRENCY",
  STOCK = "STOCK",
}

export enum BookingType {
  DEPOSIT = "DEPOSIT",
  CHARGE = "CHARGE",
  INCOME = "INCOME",
  EXPENSE = "EXPENSE",
  APPRECIATION = "APPRECIATION",
  DEPRECIATION = "DEPRECIATION",
}

export enum PeriodType {
  MONTH = "MONTH",
  QUARTER = "QUARTER",
  YEAR = "YEAR",
}